requirements.txt
agently>=4.0.8.1
PyYAML>=6.0.2
ddgs>=9.10.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
